OTM: Separate apps into their own file
