+++ v3/otm_apps.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace lgr {
+
+bool otm_j2_nu_zero_patch_test();
+
+}
