+++ v3/otm_meshless.hpp
@@ -26,5 +26,4 @@ void otm_time_integrator_step(input const& in, state& s);
 void otm_update_material_state(input const& in, state& s, material_index const material);
 void otm_run(input const& in, state& s);
 void otm_debug_output(state& s);
-bool otm_j2_nu_zero_patch_test();
 }
