+++ v3/unit_tests/mechanics.cpp
@@ -6,6 +6,7 @@
 #include <hpc_execution.hpp>
 #include <hpc_functional.hpp>
 #include <hpc_macros.hpp>
+#include <hpc_math.hpp>
 #include <hpc_matrix3x3.hpp>
 #include <hpc_numeric.hpp>
 #include <hpc_range.hpp>
@@ -14,9 +15,9 @@
 #include <lgr_input.hpp>
 #include <lgr_mesh_indices.hpp>
 #include <lgr_state.hpp>
+#include <otm_apps.hpp>
 #include <otm_meshless.hpp>
 #include <unit_tests/otm_unit_mesh.hpp>
-#include <hpc_math.hpp>
 
 namespace lgr_unit {
